tests/wpt/web-platform-tests/portals/portals-repeated-activate.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test(t => {
    assert_implements("HTMLPortalElement" in self);
    let win = window.open("resources/portal-repeated-activate-window.html");
    win.onload = () => win.activate();
    window.onmessage = t.step_func_done(() => {});
  }, "test activation in page that has been reactivated")
</script>
